+++ deployment/updates/scripts/add_permissions/2023_05_29_add_sessionCuePoint_permission.php
@@ -1,0 +1,9 @@
+<?php
+/**
+ * @package deployment
+ * Add permissions to add sessionCuePoint
+ */
+
+$script = realpath(dirname(__FILE__) . '/../../../../') . '/alpha/scripts/utils/permissions/addPermissionsAndItems.php';
+$config = realpath(dirname(__FILE__)) . '/../../../permissions/object.KalturaSessionCuePoint.ini';
+passthru("php $script $config");